requirements.txt
streamlit
pandas
plotly
pymysql
python-dotenv
